Added a Dockerfile to help with using snyk-to-html in pipelines
Packaging snyk-to-html as a Docker image makes using the tool in places
where installing a full Node toolchain is problematic or annoying. The
image also aims to make usage easier by outputting the html on stdout,
rather than writing a file to disk.

Using the -o and -i options the image can be used to read and/or write from
files on disk as well.