fix: container new json format
The `snyk container test` command has a new JSON format as discussed in
[this Notion page](https://www.notion.so/snyk/Snyk-CLI-JSON-output-for-container-projects-with-application-vulnerabilities-eb53f013f2eb47dab2aad7ca6d9a8fdc#4a32dcde34fc40a584e62c6a07c61ddb).

This means that when app vulnerabilities are enabled through the
`--app-vulns` flag (soon to be set by default as well), the JSON
document will contain an additional `applications` key that includes
information about the apps in a container plus their vulnerability
information.

This commit makes sure that these application-specific vulnerabilities
are also shown in the HTML generated by `snyk-to-html`.